Make invite URL validator Node 20 compatible
